+++ backend/src/api/__init__.py
@@ -4,6 +4,7 @@ from flask_restx import Api
 from .users import ns as user_ns
 from .scans import ns as scans_ns
 from .stats import ns as stats_ns
+from .leaderboard import ns as leaderboard_ns
 
 api = Api(
     title='Asso card API',
@@ -14,3 +15,4 @@ api = Api(
 api.add_namespace(user_ns, path="/api")
 api.add_namespace(scans_ns, path="/api")
 api.add_namespace(stats_ns, path="/api")
+api.add_namespace(leaderboard_ns, path="/api")