+++ backend/src/api/users.py
@@ -18,7 +18,13 @@ class Users(Resource):
         with Database(auto_commit=True) as db:
             result = db.query(User).all()
             for user in result:
-                user_list.append(user.to_dict())
+                user_list.append(user.to_public_dict(
+                    User.user_id,
+                    User.name,
+                    User.first_name,
+                    User.email,
+                    User.creation_date
+                ))
         return jsonify(user_list)
 
 
@@ -28,35 +34,9 @@ USER_UPDATE_PARAMS = ns.model('Updating user parameter', {
     "email": fields.String(required=True)
 })
 
-@ns.route('/user/user_id/<user_id>')
-class SingleUser(Resource):
-    @authentification_required
-    def get(self, user_id):
-        user = None
-
-        with Database(auto_commit=True) as db:
-            user = db.query(User).filter_by(user_id=user_id).first()
-            if user:
-                user = user.to_dict()
-
-        abort_if_doesnt_exist(
-            user, 
-            code=400, 
-            message="No user found with this user id"
-        )
-
-        return jsonify(user)
-
-    @ns.expect(USER_UPDATE_PARAMS, validate=True)
-    @authentification_required
-    def put(self, user_id, **kwargs):
-        return update_user(
-            card_uid=None,
-            user_id=user_id,
-            name=str(ns.payload["name"]),
-            first_name=str(ns.payload["firstName"]),
-            email=str(ns.payload["email"])
-        )
+USER_DEPT_UPDATE_PARAMS = ns.model('', {
+    "debt_update_amount": fields.Integer(required=True)
+})
 
 @ns.route('/user/card_uid/<card_uid>')
 class SingleUserCardUID(Resource):
@@ -67,7 +47,16 @@ class SingleUserCardUID(Resource):
         with Database(auto_commit=True) as db:
             user = db.query(User).filter_by(card_uid=card_uid).first()
             if user:
-                user = user.to_dict()
+                user = user.to_public_dict(
+                    User.user_id,
+                    User.name,
+                    User.first_name,
+                    User.email,
+                    User.creation_date,
+                    User.currency_amount,
+                    User.admin,
+                    User.debt_amount
+                )
 
         abort_if_doesnt_exist(
             user, 
@@ -80,17 +69,9 @@ class SingleUserCardUID(Resource):
     @ns.expect(USER_UPDATE_PARAMS, validate=True)
     @authentification_required
     def put(self, card_uid, **kwargs):
-        return update_user(
-            card_uid=card_uid,
-            user_id=None,
-            name=str(ns.payload["name"]),
-            first_name=str(ns.payload["firstName"]),
-            email=str(ns.payload["email"])
-        )
-
-def update_user(card_uid, user_id, name, first_name, email):
-
-    with Database(auto_commit=True) as db:
+        name=str(ns.payload["name"])
+        first_name=str(ns.payload["firstName"])
+        email=str(ns.payload["email"])
 
         abort_if_doesnt_exist(
             name, 
@@ -100,14 +81,23 @@ def update_user(card_uid, user_id, name, first_name, email):
             code=500
         )
 
-        user = db.query(User).filter_by(card_uid=card_uid).first() if card_uid \
-            else db.query(User).filter_by(user_id=user_id).first()
-        abort_if_doesnt_exist(user, code=400, message="No user found with this card id")
+        with Database(auto_commit=True) as db:
+            user = db.query(User).filter_by(card_uid=card_uid).first()
+            abort_if_doesnt_exist(user, code=400, message="No user found with this card id")
 
-        user.update_date = datetime.now()
-        user.name = name
-        user.first_name = first_name
-        user.email = email
-        db.commit()
+            user.update_date = datetime.now()
+            user.name = name
+            user.first_name = first_name
+            user.email = email
+            db.commit()
+        
+        return True
 
-    return True
+    @ns.expect(USER_DEPT_UPDATE_PARAMS, validate=True)
+    def post(self, card_uid):
+        debt_amount = None;
+        with Database(auto_commit=True) as db:
+            user = db.query(User).filter_by(card_uid=card_uid).first()
+            user.debt_amount += int(ns.payload["debt_update_amount"])
+            debt_amount = user.debt_amount;
+        return debt_amount;