+++ backend/src/database/scan.py
@@ -3,8 +3,9 @@ from sqlalchemy.sql.sqltypes import Date, DateTime
 from sqlalchemy_serializer import SerializerMixin
 from database.database import Base
 from datetime import datetime
+from .public_data_filter import PublicDataFilter
 
-class Scan(Base, SerializerMixin):
+class Scan(Base, SerializerMixin, PublicDataFilter):
     __tablename__ = 'scans'
     id = Column(Integer, primary_key=True)
     date = Column(DateTime, unique=False, default=datetime.now)