+++ backend/src/database/user.py
@@ -3,8 +3,9 @@ from sqlalchemy.sql.sqltypes import DateTime, Float
 from database.database import Base
 from sqlalchemy_serializer import SerializerMixin
 from datetime import datetime
+from .public_data_filter import PublicDataFilter
 
-class User(Base, SerializerMixin):
+class User(Base, SerializerMixin, PublicDataFilter):
     __tablename__ = 'users'
     user_id = Column(Integer, primary_key=True, autoincrement=True)
     card_uid = Column(String, nullable=False)
@@ -14,6 +15,8 @@ class User(Base, SerializerMixin):
     name = Column(String, default=None)
     first_name = Column(String, default=None)
     email = Column(String, default=None)
+    admin = Column(Integer, default=False)
+    debt_amount = Column(Integer, default=0)
 
     def __init__(self, card_uid, currency_amount):
         self.card_uid = card_uid