+++ hardware/src/app.py
@@ -14,7 +14,7 @@ def start_headless():
 def start_with_client():
     log("[ Client mode ]")
     eel.start(
-        'index.html?reversed=true',
+        'index.html?raspberry=true',
         mode='chrome',
         host='0.0.0.0',
         cmdline_args=[
@@ -22,7 +22,12 @@ def start_with_client():
             '--window-position=0,0',
             '--start-fullscreen',
             '--disable-features=Translate',
-            '--kiosk'
+            '--incognito',
+            '--kiosk',
+            '--fast',
+            '--fast-start',
+            '--disk-cache-dir=/dev/null'
+            '--noerrdialogs'
         ]
     )