+++ hardware/src/web/js/index.js
@@ -1,0 +1,4 @@
+window.onload = async function () {
+	const image = window.navigator.onLine ? "img/wifion.png" : "img/wifioff.png";
+	document.getElementById("connectionStatus").src = image;
+}