+++ hardware/src/web/js/transac.js
@@ -5,7 +5,17 @@ window.onload = function () {
 
 	let reason = undefined;
 	if (urlData['reason'] !== undefined)
-		reason = urlData['reason'][0];
+		switch(urlData['reason'][0]) {
+			default:
+				reason = "Erreur inconnue";
+				break;
+			case 'NEW_USER_CANT_SPEND':
+				reason = "Impossible de dépenser en tant que nouvel utilisateur";
+				break;
+			case 'NOT_ENOUGH_CURRENCY':
+				reason = "La solde du compte n'est pas suffisante";
+				break;
+		};
 
 	document.getElementById('accountMoney').innerHTML = getPriceString(price);
 	document.getElementById('clientNum').innerHTML = userId;
@@ -13,3 +23,6 @@ window.onload = function () {
 	if (document.getElementById('reasonBug'))
 		document.getElementById('reasonBug').innerHTML = reason;
 };
+
+if (parseURLParams(window.location.href)['raspberry'] != undefined)
+	document.getElementsByTagName('html')[0].style = 'transform: rotate(180deg);'